+++ angular_compiler/lib/cli.dart
@@ -13,4 +13,4 @@ import 'package:meta/meta.dart';
 export 'src/cli/builder.dart';
 export 'src/cli/flags.dart';
 export 'src/cli/logging.dart';
-export 'src/cli/messages.dart' show messages;
+export 'src/cli/messages.dart' show messages, SourceSpanMessageTuple;